Add feature_metrics_dashboard table
The feature metrics dashboard in the Support interface is slow to load
(~1.5 minutes). We want to speed this up by caching its data in a
database table.

This commit adds the migration for the new table. It uses a JSONB column
for the metrics - this gives us a flexible schema which makes it easier
to iterate on the contents of the dashboard.