3393 Add course summary component
trello-ticket: https://trello.com/c/sHdPhNdl/3393-implement-course-details-box-on-make-offer-page-in-manage

Add a new source summary component to surface the course name, provider,
location and study mode on the make a new offer page.

We cannot reuse the existing summary component since there are extra
margins and styles included, this also allows us to test the details
of the course in isolation.