Find changed apply again applications
Add `FindChangedApplyAgainApplications` service. This can be used to
count the number of candidates that have applied again and changed their
applications in a significant way. For our purposes significant means
any of the following:

- Updated `becoming_a_teacher` (personal statements)
- Updated `subject_knowledge`
- Updated/added/deleted any qualifications

We just use `as_json` to extract the attributes and associations
that are considered significant when changed. This makes the definition
of those 'significant things' more declarative and easier to change.