Unset headers when making a POST request in tests
Setting a page driver header in this test was causing other tests to fail as
all forms were expected to send requests with the content type `application/json`.

Unsetting the content type header seems to resolve the issue at the end of the test.
I also removed setting the Capybara session as it didn't seem required.