Begin refactoring ApplyFromFindController
This controller and the ApplyFromFindPage split up the logic about what
should happen to the user for different combinations of course, openness
on apply and the state of the current candidate. Push logic from the
show action into ApplyFromFindPage and make some of the naming more
explicit. The params are still mixed up - to fix later!