Fix a bug when trying to add a course from the wrong cycle
candidate_has_application_in_wrong_cycle? guards against candidates with
applications from last cycle adding course choices from this cycle, as
that's not allowed. In the event the course had come from the API, we
threw an exception on `course.recruitment_cycle_year` because that
method doesn't exist on those objects.

Because we only ever retrieve courses for the current cycle, we can
assume that if the application form isn't in this cycle, then it's not
suitable for adding this course.