Change messaging when candidate adds already chosen course
We have a validation in place which ensures that a candidate can't add
the same course to their application twice.

Rather than style this as an error message in the UI, handle the
situation more gracefully by taking the candidate to the course choices
review page and display an info banner informing them they've already
added that course.