Redirect to the interviews index if the store has been cleared
Clicking back after creating or editing an interview takes the user to the 'check' step of the interviews wizard.
The store will be cleared on creating or editing the interview so we can use this as an indication that there's nothing to check and that the user should be redirected to the interviews index.