Revert application choices updated_at for stale applications
Quite a few old applications were touched as part of a data cleanup.
When second nationality was updated it touched all the associated
application choices which has triggered them to be passed through the
API.

This PR reverts the application choices updated_at to the previous
audits created_at time. It will only do this if the latest audit
on the application form is the one that updated second_nationality
from '' to nil.