3499 Move API authentication into concern from DataAPI
trello-ticket: https://trello.com/c/yo8jZhLa/3499-spike-bare-bones-api-for-register

To allow reuse of API authentication move into a controller
concern and include in the DataAPI controller.

Since this is a concern add a new `api_type` attribute to data
records, which allows use to call on the API module when finding
the API users.