3449 Update preferences service
trello ticket: https://trello.com/c/Wm7NAOrk/3449-migrate-provider-user-notification-settings-to-new-structure

To cater for both notification worlds, allow the service to either set
the `send_notification` on a provider user and upsert the new notificaion
preferences values. This method will be used when creating a provider user
or updating the send notification values.

In the new notification world, we should also cater for the old
in the new parameters and setting both the new notifications and
old notifications. To set the old notification, if all new notifications
are the same value, the old one should be the same. If the new values
are both true and false default to true as the user has opted in to
some values.

Also add a transient setting to the provider_user factory to allow
the optional creation of the new notification preferences