Move 'course already added' messaging into course selection controller
This logical check and its accompanying error message aren't used in the
same way as the other validations in PickCourseForm. Rather, it's a
special case where we redirect with an info message in the event of a
repeat submission of a stale course selection form. Also, we want to
allow editing of a course choice without encountering this check.

Motivated by the above, do the following:

- Move the check and its messaging out of the form object and into the
  course selection controller
- Make execution of the check conditional on the presence of the
  course_choice_id param (which is only present when editing an existing
  choice)