Fix 'too many course choices' error handling
When submitting a form that creates a course option, we check if the
candidate already has the maximum number of choices. If so, we redirect
them to the application form page and show them an error message. The
error message currently has a bug where it doesn't display the course
that was selected.

Change this feature as follows:

- Correct the locale string syntax so that the selected course is
  correctly displayed in the error
- Redirect to the course choices review page instead of the application
  form page, as this is more useful for adjusting chosen courses.