Simplify initialization of PickSiteForm and AddOrUpdateCourseChoice
We can remove the following arguments from the PickSiteForm initialize
method:

- course_id
- study_mode
- provider_id

We do this as follows:

- Make `available_sites` a class method that takes the course_id and
  study_mode as arguments, allowing it to be used wherever needed
  without initializing an instance of the form. This allows us to remove
  course_id and study_mode from the arguments list.
- We can then remove the `course` and `provider` methods, the existence
  of the latter appearing to be a hangover from when we were using codes
  instead of database ids and needed to scope course retrieval by
  provider.
- With no need to retrieve a provider from the database, we can remove
  provider_id from the arguments list.

A side effect of this is that the initialize method for the
AddOrUpdateCourseChoice service can also be simplified. We still need
access to a course object inside the service, but this can be retrieved
via the course option.