Account for KeyError being raised by ActiveSupport::TimeZone
It's possible to supply a value which is parsed internally by ActiveSupport::TimeZone as a Date but doesn't contain the expected fields.
A KeyError is raised in this case, which is possibly due to over-optimistic expectations in the ActiveSupport::TimeZone#iso8601 method.
This commit rescues from this type of error to gracefully handle this invalid parameter format.