When no step is specified return the step before the last
This enables handling actions other than new and edit, where we don’t want to 
duplicate the step entry in the wizard path history by having to configure the step and have no knowledge of the view action e.g. new, edit, as those are only posted when the back button is clicked.