Ensure the date and time passed to TimeLimitCalculator has timezone information
ActiveRecord 'maximum' returns the max value as a Time (without timezone). When passed to TimeLimitCalculator the results are also without timezone information. The time zone appears to be applied on AR update and so the offset is added. For end of day times this skews the value to the following day at 00:59am.
We have seen this happen for decline by default date times around the summer time BST adjustments.
This commit ensures that when setting the decline by default date we pass the TimeLimitCalculator a time with zone information so that the results we get back also have zone information.
Also removes the spec instance double of TimeLimitCalculator as this was always returning a TimeWithZone which made replicating the issue difficult.