Rails 6.1 upgrade - change how we use ActiveSupport::Duration
We retrieve "parts" of a duration when formatting a number of months
into a years and months string. Pre-upgrade, the `parts` method returned
a Hash with a default value of 0. It now returns a Hash with the more
typical behaviour of returning nil for missing keys.

Update our `format_months_to_years_and_months` helper to reflect this
change.