Change how references are retrieved in CandidateJourneyTracker
The "completed references" fields of this export are supposed to show
timestamps for the first and second reference received for a given
application. Currently the export retrieves these references ordered by
id, irrespective of when they were actually received.

Further, some references in the production database have blank
feedback_provided_at timestamps, leading to blank fields in the export.

Address both of these issues as follows:

- Update the reference retrieval logic so that all `feedback_provided`
  references are returned ordered by their `feedback_provided_at`
  timestamp
- If a `feedback_provided` reference has a missing
  `feedback_provided_at` timestamp, rather than show an incomplete and
  potentially confusing/misleading picture in the export, opt instead
  to not display any reference data for that application. This allows
  the application to be filtered out of any reference-related analysis.