Destroy blank application when prefilling candidate application form
- Prefilling a blank application form involves creating a new
application form with data and adding this to a candidate's application
forms. When a prefilled application is created, a random creation date
is assigned which can cause the associated system spec to fail because
the blank application form may potentially have a creation date that is
later than that of the prefilled application, AND a candidate's current
application is now determined by ordering all application forms by
created_at.
- This change ensures that blank application forms are deleted before a
prefilled application is assigned to a candidate