Rails 6.1 upgrade - change ParameterMissing handling in API controllers
The message for this exception is now multiline. In addition to the
expected error message, it now also contains a 'Did you mean?' section
offering suggestions for alternative keys.

Update our handling for this error to maintain the existing behaviour
(ie - only return the main error message in the first line).