Add CSP exceptions needed for app to run
We pull in a handful of scripts from external sites, for ZenDesk and
Google Analytics. We can whitelist these.

We can use a nonce to validate inline JavaScript.

Also permit unsafe_inline for inline CSS. There are a few places where
we have inline CSS. This isn't ideal but given that some of the inline
CSS is pulled in from dependencies it doesn't appear to be possible to
use the same nonce technique as we use for JavaScript.

ZenDesk integration also uses Web sockets (for live chat I guess) and
this too can be whitelisted.