Fix validation error when arriving from find with course params
At the moment, if a candidate has an apply_2 application form
with an application choice, when they arrive from Find
with course params then they are told that they cannot
add more than 3 courses assoicated with their application form.

This commit checks to see if the application form is apply_1
or apply_2 before rendering the correct error message.