Extract DataExportHistoryComponent
This can be used on the existing export history page (for each type) as
well as the new global history page.

For the existing single-type history page we don't need to show the name
column. (The name is already in the title of the page and is the same
for all entries).