Add review functionality to personal statement section
We want to encourage candidates to review feedback from past,
unsuccessful applications. This commit adds a review workflow to the
'Why do you want to teach' section of the personal statement.

This new workflow is implemented as a set of cosmetic changes on top of
the existing 'completed' flags we have for each section. This means we
can use the existing database column and preserve the same validation
behaviour while changing only the messaging the candidate sees.

Broadly, this change does the following:

- Mark the 'becoming_a_teacher' section as incomplete when applying
  again if there is structured feedback for this section in the
  previous application form.
- Show a 'Review' badge for this section on the application form page.
- Add a RejectionReasonsHistory class that handles retrieval of past
  rejection reasons, providing methods to return the entire rejection
  history of a section or just the most recent one.
- Summarize the rejection history of the section when reviewing it.