Show pending review error on application submit
When submitting an application, we show error messages for any
incomplete sections. If the section is one which is pending review, we
want to show slightly different messaging to the candidate.

Apply the changes required to make this happen for the
'becoming_a_teacher' section, the only section thus far for which the
review workflow is implemented.