Restoring the business day logic for dbd specs
It seems that the logic for calculating dbd business days changes
if it is outside hours. Substracting 12 hours was instated to ensure we
don't roll over to next day.

I removed this initially as it was failing, however I've tested again
and now it seems ok with subtracting hours and adding hours.