Include accredited providers as part of check for candidate masquerading
If a provider user belongs to an org which ratifies a course a candidate has applied for, they should be able to masquerade as the candidate (regardless of permissions).
This commit widens the check to verify that the provider user can also belong to an accrediting org from the candidate's applications in order to masquerade.