Remove DBD specs about running the service twice
These tests are flaky, and they are also misleading.

On a bank holiday weekend, the end of the next business day from
Saturday is the end of Tuesday.

If you make the same calculation on Sunday, the next business day will
also be the end of Tuesday.

i.e. changing the date from which you calculate DBD will not always give
you a different DBD

Therefore adding days to the last_decision_at and recalculating DBD is
dependent on context, and these specs run in "real" time.

I think the behaviour of changing DBD if a later timestamp could be
reasonably expected from SetDeclineByDefault, and the logic is no
different.