Reinstate offer form and controller
As part of the new reinstate offer flow a controller and a new
form have been added to handle loading and updating the status of
the candidates course choice. The form will also validate that the
support user has confirmed that they've read the guidance and that
they have also provided a valid zendesk ticket link.