Add the ability to search Vendor API valdiation errors
To build the search page for Vendor API validation errors, we need
the ability to pass different parameters and search Vendor API requests.

To search for specific errors in the response body, we utilise the
jsonb query format to allow searching multiple errors nested in the
same response body.

For the candidate search ability, we had a nested class in the controller,
move this out into the validation error class as it is more suited there, also
add missing testing coverage.