Remove candidate sign-in confirmation email
It's only possible to sign in via a magic link, which is delivered to
the candidate via email. If an attacker controls the candidate's email
account, they're also going to see the sign-in confirmation email
and be able to delete it.

Therefore it's not useful to send this to candidates.