Add query string as well as request body to search view for errors
If a request is a POST, query strings are not included in the
request body. To surface the whole picture for validations errors,
add a new section for query strings as well as request bodies in the
search error view.

If either of them is not populated we render the empty value, otherwise
we show the formatted content.