Make missing section messaging consistent on review screen
When reviewing the application just prior to submit, we show info
banners for each section that hasn't been marked as complete. The
messaging in these banners is inconsistent, with some referring to
details that haven't been "entered".

This commit makes the messaging consistently refer to the section as
having not been "marked as complete".