Identify and deduplicate bad self-ratifying provider relationships
There are 9 provider relationships in production where the training and
ratifying provider have the same name. There are courses for each of
these providers which are provided/ratified by the two sides of the
relationships. This seems to be bad data as a self-ratified course
should have a nil accredited provider and the duplication of provider
records and relationship records are unnecessary.

This commit finds these relationships and iterates over courses
concerned nillifying the accredited_provider and making them
self-ratified. The relationship and duplicate provider records are then
destroyed.