Don't show withdrawal reasons if choice is not 'offer_withdrawn'
In a couple of places in the app, we have components which summarise a
candidate's course choices. In the scenario where an 'offer_withdrawn'
application choice has another offer made against it, we still show the
withdrawal reasons in the summary.

Hide the reasons in this scenario by checking the state of the
application choice.