Do not use JavaScript for back links
We can't use javascript: URLs because they violate CSP. We can't set
unsafe-inline (to get around CSP) as we already have nonces (see parent
commit).

If we lack a referer for a back link, fall back to the service_link (ie
the respective start page of the current interface). Don't try to get
clever with the JS History API, which can do weird things like send you
to a different website altogether.

This means we don't ever need to hide back links, so we can remove the
--no-js styling too.