Update Referee Type and Name Forms and Controllers
In order to minimise the number of blank fields displayed to candidates on the review page when they add an
incomplete reference. Altering the type controller to pass the type through to the name controller as a URL param.
Altering the name controller to submit and create the reference with these two values. Updating form methods to represent
this