Fix for flakey time-dependent test
This test failed late in the evening with an off by one 1 day error when
asserting a date output. I think this is because the test takes a
finite amount of time to complete it's steps and I got unlucky with the
assertion happening the day after the code under test ran. So a bit of
an edge case. I think the way to fix this is with `Timecop.freeze`
rather than `travel`.