Use the method append_info_to_payload for logging
In the previous loggin methods, we could set and get RequestLocal store
values. However due to the way the new logger is initialised the RequestLocal
current store id isn't being set in the custom logger.

We can utilise the method `append_info_to_payload` to capture logs from
our different controllers and customise the parameters we would like to
surface. Ideally this would live in the application controller,
however we aren't using that so we will make do of this in the relevant
controllers.