Configure seconds value to BusinessTime end of day
The end of the business day was set to 23:59 and the end_of_day date helper uses the time 23:59:59.999999999.
These tests were failing when the revised dbd date fell at the end of
the week or weekend ie. after the close of business. This meant depending on which day
the test was run we'd see a failure because the rails helper calculates
the end of day just outside the business day and any further business
day calculations make the equality comparison off by one day.

This commit adds the seconds value for the end of the business day in
BusinessTime config.