Link back to personal details review page when editing
This changes the back link so that it no longer relies upon the HTTP referer header when editing an answer to a question.

This fixes a bug where if you edit a question, and then get a validation error and get redirected back to the same page, the back link links to the page you're already on.