Add data migration to fix ethnicity_background typo and add hesa codes
There was a typo in an app helper which supplied ethnicity background values, this lead to a failure in finding the correct hesa_ethnicity code for 'Caribbean' ethnicity_background.
This data migration finds these records (there are currently 88 in production) and fixes the typo and adds the relevant HESA code ('21').
The typo in the helper has been fixed in the previous commit.