Fix bug on Previous Unsubmitted Application Review
A candidate has encountered a bug when clicking through to the previous
application review view. On Investigation this is because they carried
over their application as opposed to applying again. As submitted_at is
nil in this case it breaks a helper method. This PR updates this method
to return nil before hitting that line in this case.