Allow more than two references to be provided
We're implementing a feature that allows candidates to select which
references get submitted as part of their application (as opposed to
automatically using the first two that receive feedback).

As part of this, we need to allow the candidate to acquire more than two
references.

With the 'reference_selection' feature flag active:

- Don't cancel outstanding references when two references are received
- Don't show errors about having too many references
- Update the ReferenceActionsPolicy to ignore received reference count
  when deciding on permitted actions

Also included in this commit is a system spec outlining the new
behaviour.

Co-authored-by: James Glenn <47917431+JR-G@users.noreply.github.com>