Add section complete functionality to references section
We're implementing a feature that allows candidates to select which
references get submitted as part of their application (as opposed to
automatically using the first two that receive feedback). This has
implications for how the references section gets marked complete.

Currently there is no explicit database flag for reference section
completion - the candidate simply sees the section marked complete in
the UI once two references are provided.

Change this so that (conditional on the reference_selection feature
flag), we have section complete radio buttons like other parts of the
application.