Fix TT API incremental sync when no last_sync present
If an incremental sync is requested in an environment that hasn't yet
populated a last_sync entry, course sync breaks with a nil error.

Fix by checking last_sync is present before updating the database scope
with an updated_since.

Encountered when running the 'setup_local_dev_data' rake task.