Fix test failure in `DisabilityStatusForm`
This is failing when run locally, though it seems to have passed CI
(though I don't fully understand how).

The change here is to skip resetting `hesa_disabilities` if it's already
`nil` or some other blank value. It's just not necessary to replace
`nil` with `[]` so the original code was a bit over-zealous. We only
ever need to reset values where there are one or more HESA disability
codes present.