Temporarily disable reference_selection in the test suite
A large number of tests are breaking as we build out the initial parts
of this feature. For example, there is some coupling between the logical
checks made when requesting/receiving references and when submitting the
application (the latter of which hasn't been adapted for reference
selection).

For now, disable the feature flag across the entire test suite,
selectively switching it on wherever we test the new functionality.

This approach allows us to build out gradually without having to
keep the entire suite passing with each step. This line of config
will be removed prior to going live with the feature, most likely
once the full reference selection flow (with app submission) is
implemented.