Update all the backlinks in the references section to have explicit backlinks
All the backlink related behaviour here is a bit of a nightmare to
follow. This commit fixes issues where a candidate can click to create a
reference, choose a type and name, therefore persisting the reference,
then clicks back multiple times and picks a new type.

Previously, once they picked a new type and input a new name, they would
have two references as the type and id were not being passed through the
backlinks.