Ensure that candidates can only edit non-requested references
As id is now an optional param to the new actions for type and
name, we need to guard against someone manually visiting the
endpoint and updating these attrs.

This commit ensures that the reference is editable
if a candidate visits the new actions with an id in the query string.