Set dbd day to last descision timestamp
We are setting the DBD through business days from now. However due to
weekends the DBD can be true to a certain date for 3 days:
Saturday/Sunday/Monday. This however causes failures with us
setting the `last_decision_at` to two days before as it doesn't
take into account weekends.

If we set the DBD to be relative to `last_decision_at` we no longer
have the dependency on weekends and thus resolve the flakey tests.
This has been tested with the next 100 days in the future and there were
no failures.