Add pre-translated error messages
ActiveModel errors.messages will call Error.generate_message and attempt to interpolate
the translation which is proving costly. Adding the error messages as
Strings bypasses this mechanism.

It's likely that there's either too much recursion in the validation
here or that the many calls to @wizard.errors.messages are not
performant, either way this reduces a very long response time when all
fields fail validation.