Add option to SelectedReferencesComponent for incomplete banner
As with other review components, add a `show_incomplete` option to the
selected references one.

This allows us to show the incomplete banner on the application review
page, but ensures it isn't rendered by default when all we want to do is
display the selected references table (ie - on the reference selection
review page).

Alongside this change, rework the specs for the component and the
reference selection app submit system spec, making the testing slightly
more thorough.