Push sync updated_since two hours in the past
We currently persist an updated_at timestamp in Redis to help with our
incremental syncs of the TTAPI.

However the teacher training API is 1 hour behind. Since we kick off
subsequent jobs for syncing providers then courses then sites, and we
reset the updated_at after syncing providers, we were never
using the correct timestamp to query for courses, and always missing
the 10 minute window.

Set the updated_at timestamp to 2 hours to avoid issue for now