Ensure we set '00' hesa_disabilities code
If the candidate has already answered some of the equality and diversity questions but answers 'no' to disability status we still need to populate the appropriate '00' HESA code.
The code was previously expecting the equality_and_diversity field to be nil.

Similarly if a candidate changes their answer to 'no' for an application
with existing disabilities data we should set the appropriate '00' code.