Add data migration to remove surplus reference selections
To prepare Apply for the reference selection feature, a prior data
migration (a1ca03f5fb) set `selected: true` for all `feedback_provided`
references. Prior to the introduction of reference selection, we
expected there to be a maximum of 2 provided references per application.
Historic bugs in the service, however, mean that we have some
applications (90 in total) with more than 2 `feedback_provided`
references.

This means that those references are now also selected. It would be
useful to have an invariant check that warns us about surplus selections on
an ongoing basis. This data migration removes existing surplus
selections in preparation for such a check.

The migration:

- Looks for all applications with a selected reference count of > 2
- Sets `selected: false` on all surplus selections
- Does not care particularly about which references get de-selected
  (which shouldn't be an issue as this migration will be run prior to
  the reference selection feature going live)