Only show orgnisations list for set up permissions in account
When a provider navigates to their account, and the organisation
section, only surface the list of organisations with permissions
with `setup_at` that is not `nil`.

However, if the provider user is associated to a provider that is
also a ratifying provider, but they cannot manage it, surface the
permission. This is because we currently always surface those permissions
as read only, with dedicated content for ones not currently set up.