Don't rely on the ordering of providers from NotificationsList#for
Instead of testing the first and second recipients of emails sent by the service we should just verify that both the training provider and ratifying provider have received an email.
NotificationsList#for does not contain an ordering clause so we can't be sure which provider user will be returned first.