Sort permissions consistently
The order in which training and ratifying permissions are returned is not consistent. The queries include the training and ratifying providers so sort on the provider name on the 'other' side of the organisation in question ie. for training provider permissions, sort on ratifying provider name and for ratifying provider permissions sort on the training provider name. This ensures the relationships maintain a predictable order when updating multiple relationships.