Fix interview system spec filling in dates
When changing dates in the interview test, we are setting the day forward but
not the month. If this happens to be at the end of the month, this will cause
the test to fail as we are changing the date for instance from the 30th
of June to the 1st of June without changing the month.

Make sure to change all the values for the day, month, year when setting
the new date.