Simplify decline by default system spec
At some point in the past, the copy in the decline by default email was
simplified so that it did not vary depending on whether or not the
candidate had a rejected application choice.

Update the steps in the system spec so we have just the one method that
asserts the contents of the received email.

Further, the content of the email now dynamically changes depending on
where we are in the cycle. As that isn't the focus of this system spec,
we look for a more generic piece of content that exists irrespective of
cycle date.