Revert "Allow monitoring the staging app"
This reverts commit d32cd10290ee859026a0676b899cdb15ef8425cd.

The deployment failed with error:
Error: Element not found

  on modules/paas/network-policies.tf line 17, in data "cloudfoundry_app" "prometheus_app":
  17: data "cloudfoundry_app" "prometheus_app" ***

This is because the network policy is created by the staging paas user
but it needs access to the production space as prometheus is in
production. We can't let this happen to keep environments segregated.

This may need a refactor if it is required.