Enable candidates to select site when one study mode is full
If a candidate wishes to apply for a course which is advertised as full
or part time, and one of the study modes has no vacancies, they can't
apply for the course at all because they're unable to choose a site.

This commit fixes that behaviour by picking the remaining available
study mode when we move to the site step.

More context:

https://ukgovernmentdfe.slack.com/archives/CP18YJXPY/p1626296273211900