Add an authentication shortcut for load testing accounts
When load testing Apply, test users driven by JMeter will create
accounts. We need some way to sign in these users via the magic link
that gets emailed out.

Introduce logic to the SignUpController that checks for a load testing
email address, and if found redirects the candidate to the magic link.