Roll the year forward when using the cycle switcher
The current_year method does not take into account the cycle switcher
when testing locally, so the application_form.recruitment_cycle_year
is always a year ahead of the current_year. The result of this is that
you can't add courses when selecting Find/Apply opens on the cycle
switcher.