Use `find_by` to avoid error on offerable courses
If courses are not offerable and we try to find available course
options or study modes, we are returning a 404 in manage because the
query is erroring on `find` as the course can't be found.
Use `find_by` instead to return an empty list and avoid erroring/returning
a 404 from the controller.