Handle users belonging to both orgs in a relationship
When setting up org permissions we need to prefer an organisation for users who belong to both orgs in a relationship.
In this case prefer the alphabetically first org name and mail managing users for this org.

When making changes we should always be making the change from the
context of a specific organisation so use this to infer the partner org.