Cache application form sections on unsubmitted show page
This page has heavy traffic given it's the hub for all application
edits. It also takes a fair bit of work to render the template (14 SQL
queries). There aren't any unnecessary queries as such, and it isn't
straightforward to reduce the total number.

We can reduce this query count to 1 by caching most of the application
form sections. We use the updated_at columns in both the application
form and feature flag tables to create a cache key. Any changes to the
application form, or changes to feature flag state, will result in a
cache miss.