Remove unnecessary Timecop freeze from system spec
This test fails because it invokes the helper method that steps through a
candidate completing all sections of an application. It also freezes time
around this. We now cache part of the unsubmitted application form
page, and the frozen time means that no cache busting takes place. This
causes the spec to fail.

The freeze time call does not appear to affect the behaviour of the
test, so remove it.