Fragment cache the provider select box
The provider screen in the course choice flow has a high object
allocation according to Skylight.

Cache the select box with a cache key that's dependent on the most
recent updated_at in the providers table. This will retrieve that part
of the view from the cache instead of holding all providers in memory to
render it.