Prevent additional queries on qualifications
ApplicationForm#qualification_in_subject is called multiple times per row of data for an applications export.
This happens because we iterate over relevant gcse subjects in order to present a summary. We've already included
ApplicationQualifications in the main data export query, so this commit avoids queries to re-retrieve gcse and degree
qualifications and uses Ruby instead to collate the relevant information.
For a single row of data we were making an additional query to find the 'first' degree and 3 more queries to find gcses
of various subjects.