Cache providers content page
According to Skylight this page is quite slow. It has a high rating
because it's also fairly popular. One of the main issues appears to be
a high number of object allocations when rendering the page.

We're also planning on removing the page prior to the start of the next
cycle. Therefore, rather than changing its design or performance tuning
the existing code, it probably makes sense to just cache it instead.

Enclosing most of the template in a cache block with a 15 minute expiry
means that:

- We ease the performance burden of the page in the short term
- In case we don't quite get round to removing it prior to the start of
  the cycle, we reduce its impact on overall site performance during the
  expected spike in traffic