Weight sidekiq queues to distribute work in priority order
https://github.com/mperham/sidekiq/wiki/Advanced-Options
Weighting Sidekiq queues means that work is distributed according to the weight given, this means that a queue with a weight of 2 is checked for jobs twice as often as a queue with a weight of 1.
We tend to use the low_priority queue for our workflow tasks which would currently be blocked by jobs in the mailers queue because of the sequential priority of our current config.
This commit attempts to allow work on all queues to proceed while still applying priority to specific queues.