Ensure that candidates with missing GCSEs use the new flow
This commit ensures that during carry over, applicants with missing GCSE
explanations have the correct attrs set so that they can use the new flow
in the upcoming cycle.

Previously, we asked candidates to provide details of how they were
attaining their GCSE. We now use the not_completed_explanation field to
capture this and use the missing_explanation to capture their answer on
why they don't need one.

Previously we did not ask for this information so this should now be set
to nil.