Make GetCourseOptionFromCodes more lenient
This allows an API user to leave out a site code when changing an offer. If a single course option can be determined without the site, then good. Otherwise let the user know it is ambiguous. We are making this change so that vendors without site code knowledge can still integrate with the offers API