Add hidden blank input field to checkbox filters HTML
Some of the filters store their state in the session so that the options are remembered after navgating away. A side effect of this is that when going to the filter page with no URL params, the saved state is loaded rather than a blank one. Therefore, when no filters are selected, we must send that state back in the params rather than sending no state.