Always fast_forward by at least 1 day
This caused problems when we wanted to create an apply 2 test
application in the offer state. The original A1 application needed to
exist and have been rejected, but it was possible for it to have been
created_at at the same second as the A2 application receiving the offer.
This triggered an OfferValidation error: "You cannot make an offer
because you can only do so for the most recent application"