Refactor chaser, permissions and relationships checks
Reworks chaser checking logic so that if the user has received the Find
service opening email but not the permissions email they will still be
notified to remind them to complete permissions setup. The intention is
that we can run this process and pick up users who have recently been
granted manage_organisations permissions.

Simplifies how we check for pending relationships requiring setup and
whether the user is eligible to perform setup.

Works out the query limit based on the hours remaining and the number of
providers who have not yet been sent emails.